+++ lib/parlour.rb
@@ -3,6 +3,8 @@ require 'sorbet-runtime'
 
 require 'parlour/version'
 
+require 'parlour/debugging'
+
 require 'parlour/kernel_hack'
 
 require 'parlour/plugin'
@@ -19,6 +21,7 @@ require 'parlour/rbi_generator/constant'
 require 'parlour/rbi_generator/namespace'
 require 'parlour/rbi_generator/module_namespace'
 require 'parlour/rbi_generator/class_namespace'
+require 'parlour/rbi_generator/enum_class_namespace'
 require 'parlour/rbi_generator'
 
 require 'parlour/conflict_resolver'